request function moved to RegistrationView and data received passing to FifthStep set up
